+++ Specs/Renderer/AutomaticUniformSpec.js
@@ -106,6 +106,21 @@ defineSuite([
            13.0, 14.0, 15.0, 16.0));
     });
 
+    it('has czm_inverseModel', function() {
+        var fs =
+            'void main() { ' +
+            '  gl_FragColor = vec4(' +
+            '    (czm_inverseModel[0][0] == 1.0) && (czm_inverseModel[1][0] == 0.0) && (czm_inverseModel[2][0] == 0.0) && (czm_inverseModel[3][0] == -7.0) &&' +
+            '    (czm_inverseModel[0][1] == 0.0) && (czm_inverseModel[1][1] == 1.0) && (czm_inverseModel[2][1] == 0.0) && (czm_inverseModel[3][1] == -8.0) &&' +
+            '    (czm_inverseModel[0][2] == 0.0) && (czm_inverseModel[1][2] == 0.0) && (czm_inverseModel[2][2] == 1.0) && (czm_inverseModel[3][2] == -9.0)' +
+            '  ); ' +
+            '}';
+        verifyDraw(fs, new Matrix4(1.0, 0.0, 0.0, 7.0,
+            0.0, 1.0, 0.0, 8.0,
+            0.0, 0.0, 1.0, 9.0,
+            0.0, 0.0, 0.0, 1.0));
+    });
+
     it('has czm_view', function() {
         var us = context.getUniformState();
         us.setView(new Matrix4( 1.0,  2.0,  3.0,  4.0,
@@ -298,6 +313,29 @@ defineSuite([
         verifyDraw(fs);
     });
 
+    it('has czm_inverseViewProjection', function() {
+        var us = context.getUniformState();
+        us.setView(new Matrix4(1.0, 0.0, 0.0, 0.0,
+                               0.0, 1.0, 0.0, 8.0,
+                               0.0, 0.0, 1.0, 0.0,
+                               0.0, 0.0, 0.0, 1.0));
+        us.setProjection(new Matrix4(1.0, 0.0, 0.0, 0.0,
+                                     0.0, 1.0, 0.0, 0.0,
+                                     0.0, 0.0, 1.0, 9.0,
+                                     0.0, 0.0, 0.0, 1.0));
+
+        var fs =
+            'void main() { ' +
+            '  bool b0 = (czm_inverseViewProjection[0][0] == 1.0) && (czm_inverseViewProjection[1][0] == 0.0) && (czm_inverseViewProjection[2][0] == 0.0) && (czm_inverseViewProjection[3][0] ==  0.0); ' +
+            '  bool b1 = (czm_inverseViewProjection[0][1] == 0.0) && (czm_inverseViewProjection[1][1] == 1.0) && (czm_inverseViewProjection[2][1] == 0.0) && (czm_inverseViewProjection[3][1] == -8.0); ' +
+            '  bool b2 = (czm_inverseViewProjection[0][2] == 0.0) && (czm_inverseViewProjection[1][2] == 0.0) && (czm_inverseViewProjection[2][2] == 1.0) && (czm_inverseViewProjection[3][2] == -9.0); ' +
+            '  bool b3 = (czm_inverseViewProjection[0][3] == 0.0) && (czm_inverseViewProjection[1][3] == 0.0) && (czm_inverseViewProjection[2][3] == 0.0) && (czm_inverseViewProjection[3][3] ==  1.0); ' +
+            '  gl_FragColor = vec4(b0 && b1 && b2 && b3); ' +
+            '}';
+
+        verifyDraw(fs);
+    });
+
     it('has czm_modelViewProjection', function() {
         var us = context.getUniformState();
         us.setView(new Matrix4(1.0, 0.0, 0.0, 0.0,
@@ -315,6 +353,33 @@ defineSuite([
             '  bool b1 = (czm_modelViewProjection[0][1] == 0.0) && (czm_modelViewProjection[1][1] == 1.0) && (czm_modelViewProjection[2][1] == 0.0) && (czm_modelViewProjection[3][1] == 8.0); ' +
             '  bool b2 = (czm_modelViewProjection[0][2] == 0.0) && (czm_modelViewProjection[1][2] == 0.0) && (czm_modelViewProjection[2][2] == 1.0) && (czm_modelViewProjection[3][2] == 9.0); ' +
             '  bool b3 = (czm_modelViewProjection[0][3] == 0.0) && (czm_modelViewProjection[1][3] == 0.0) && (czm_modelViewProjection[2][3] == 0.0) && (czm_modelViewProjection[3][3] == 1.0); ' +
+            '  gl_FragColor = vec4(b0 && b1 && b2 && b3); ' +
+            '}';
+
+        verifyDraw(fs, new Matrix4(
+                1.0, 0.0, 0.0, 7.0,
+                0.0, 1.0, 0.0, 0.0,
+                0.0, 0.0, 1.0, 0.0,
+                0.0, 0.0, 0.0, 1.0));
+    });
+
+    it('has czm_inverseModelViewProjection', function() {
+        var us = context.getUniformState();
+        us.setView(new Matrix4(1.0, 0.0, 0.0, 0.0,
+                               0.0, 1.0, 0.0, 8.0,
+                               0.0, 0.0, 1.0, 0.0,
+                               0.0, 0.0, 0.0, 1.0));
+        us.setProjection(new Matrix4(1.0, 0.0, 0.0, 0.0,
+                                     0.0, 1.0, 0.0, 0.0,
+                                     0.0, 0.0, 1.0, 9.0,
+                                     0.0, 0.0, 0.0, 1.0));
+
+        var fs =
+            'void main() { ' +
+            '  bool b0 = (czm_inverseModelViewProjection[0][0] == 1.0) && (czm_inverseModelViewProjection[1][0] == 0.0) && (czm_inverseModelViewProjection[2][0] == 0.0) && (czm_inverseModelViewProjection[3][0] == -7.0); ' +
+            '  bool b1 = (czm_inverseModelViewProjection[0][1] == 0.0) && (czm_inverseModelViewProjection[1][1] == 1.0) && (czm_inverseModelViewProjection[2][1] == 0.0) && (czm_inverseModelViewProjection[3][1] == -8.0); ' +
+            '  bool b2 = (czm_inverseModelViewProjection[0][2] == 0.0) && (czm_inverseModelViewProjection[1][2] == 0.0) && (czm_inverseModelViewProjection[2][2] == 1.0) && (czm_inverseModelViewProjection[3][2] == -9.0); ' +
+            '  bool b3 = (czm_inverseModelViewProjection[0][3] == 0.0) && (czm_inverseModelViewProjection[1][3] == 0.0) && (czm_inverseModelViewProjection[2][3] == 0.0) && (czm_inverseModelViewProjection[3][3] ==  1.0); ' +
             '  gl_FragColor = vec4(b0 && b1 && b2 && b3); ' +
             '}';